+++ src/app/pages/demo-components/navigation-manager/navigation-manager.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { DemoNavigationManagerRoutingModule } from './navigation-manager-routing.module';
+
+@NgModule({
+  declarations: [],
+  imports: [DemoNavigationManagerRoutingModule]
+})
+export class DemoNavigationManagerModule { }